+++ tests/asgi/test_testing_asgi.py
@@ -3,7 +3,8 @@ import time
 import pytest
 
 import falcon
-from falcon import asgi, testing
+from falcon import testing
+from . import _asgi_test_app
 
 
 @pytest.mark.asyncio
@@ -106,28 +107,9 @@ def test_is_asgi_app_cls():
 
 
 def test_cookies_jar():
-    class Bar:
-        async def on_get(self, req, resp):
-            # NOTE(myusko): In the future we shouldn't change the cookie
-            #             a test depends on the input.
-            # NOTE(kgriffs): This is the only test that uses a single
-            #   cookie (vs. multiple) as input; if this input ever changes,
-            #   a separate test will need to be added to explicitly verify
-            #   this use case.
-            resp.set_cookie('has_permission', 'true')
+    client = testing.TestClient(_asgi_test_app.application)
 
-        async def on_post(self, req, resp):
-            if req.cookies['has_permission'] == 'true':
-                resp.status = falcon.HTTP_200
-            else:
-                resp.status = falcon.HTTP_403
-
-    app = asgi.App()
-    app.add_route('/async_jars', Bar())
-
-    client = testing.TestClient(app)
-
-    response_one = client.simulate_get('/async_jars')
-    response_two = client.simulate_post('/async_jars', cookies=response_one.cookies)
+    response_one = client.simulate_get('/jars')
+    response_two = client.simulate_post('/jars', cookies=response_one.cookies)
 
     assert response_two.status == falcon.HTTP_200
